Add the service layer
