+++ src/routes/profile/profile.route.ts
@@ -1,10 +1,13 @@
 import express from 'express'
-import { getProfile, updateProfile } from '../../controllers/profile.controller'
+import {
+  getProfileHandler,
+  updateProfileHandler
+} from '../../controllers/profile.controller'
 import { requireAuth } from '../../controllers/auth.controller'
 
 const profileRouter = express.Router()
 
-profileRouter.get('/profile', requireAuth, getProfile);
-profileRouter.put('/profile', requireAuth, updateProfile);
+profileRouter.get('/profile', requireAuth, getProfileHandler)
+profileRouter.put('/profile', requireAuth, updateProfileHandler)
 
 export default profileRouter
